+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Removed
 
 ### Fixed
+- i18n labels in the backoffice service view (@goreck888)
 
 ### Security